added @NSBASE@ replacement
This variable can be used in the template. It will contain the base name
of the namespace the created page is located in.

Eg. if the new page is 'foo:bar:baz:bang', @NSBASE@ will contain 'baz'

This is useful when construction data plugin tables that link to
namespaces, but use alias prefixes to address the full namespace.