Load also fields of other plugins(incl warnings for dataplugin)
nameconvention for fields from other plugins are:
<pluginname>_<componentname>
This is also the convention for the plugin component name in DokuWiki.
So fields by other plugins can only be placed in the `helper/` folder of
the plugin.

Throw (non-translated) warnings when the previous field name
'dataplugin' of the field 'data_aliastextbox' of the Data Plugin is
used.