upload files to arbitrary locations
This commit provides an addtional "file" syntax which allows us to decide where we want to store files send by bureaucracy forms. The new optional additional argument (directly after the label) is a dokuwiki namespace where the files are uploaded. We can use "." and ".." to define the namespace relative to a newly created page. By default "." is used as an file upload namespace.

The commit only affects the template action.